Extract tests for NestSingletonMethods rewriter
